Move error_code definitions to source file
