+++ hpx/error_code.hpp
@@ -107,7 +107,7 @@ namespace hpx
         ///               (if mode is \a rethrow).
         ///
         /// \throws nothing
-        inline explicit error_code(error e, throwmode mode = plain);
+        HPX_EXPORT explicit error_code(error e, throwmode mode = plain);
 
         /// Construct an object of type error_code.
         ///
@@ -124,7 +124,7 @@ namespace hpx
         ///               (if mode is \a rethrow).
         ///
         /// \throws nothing
-        inline error_code(error e, char const* func, char const* file,
+        HPX_EXPORT error_code(error e, char const* func, char const* file,
             long line, throwmode mode = plain);
 
         /// Construct an object of type error_code.
@@ -141,7 +141,7 @@ namespace hpx
         ///
         /// \throws std#bad_alloc (if allocation of a copy of
         ///         the passed string fails).
-        inline error_code(error e, char const* msg, throwmode mode = plain);
+        HPX_EXPORT error_code(error e, char const* msg, throwmode mode = plain);
 
         /// Construct an object of type error_code.
         ///
@@ -161,7 +161,7 @@ namespace hpx
         ///
         /// \throws std#bad_alloc (if allocation of a copy of
         ///         the passed string fails).
-        inline error_code(error e, char const* msg, char const* func,
+        HPX_EXPORT error_code(error e, char const* msg, char const* func,
                 char const* file, long line, throwmode mode = plain);
 
         /// Construct an object of type error_code.
@@ -178,7 +178,7 @@ namespace hpx
         ///
         /// \throws std#bad_alloc (if allocation of a copy of
         ///         the passed string fails).
-        inline error_code(error e, std::string const& msg, throwmode mode = plain);
+        HPX_EXPORT error_code(error e, std::string const& msg, throwmode mode = plain);
 
         /// Construct an object of type error_code.
         ///
@@ -198,13 +198,13 @@ namespace hpx
         ///
         /// \throws std#bad_alloc (if allocation of a copy of
         ///         the passed string fails).
-        inline error_code(error e, std::string const& msg, char const* func,
+        HPX_EXPORT error_code(error e, std::string const& msg, char const* func,
                 char const* file, long line, throwmode mode = plain);
 
         /// Return a reference to the error message stored in the hpx::error_code.
         ///
         /// \throws nothing
-        inline std::string get_message() const;
+        HPX_EXPORT std::string get_message() const;
 
         /// \brief Clear this error_code object.
         /// The postconditions of invoking this method are
@@ -219,15 +219,15 @@ namespace hpx
         ///
         /// \note This function maintains the error category of the left hand
         ///       side if the right hand side is a success code.
-        inline error_code& operator=(error_code const& rhs);
+        HPX_EXPORT error_code& operator=(error_code const& rhs);
 
     private:
         friend boost::exception_ptr detail::access_exception(error_code const&);
         friend class exception;
         friend error_code make_error_code(boost::exception_ptr const&);
 
-        inline error_code(int err, hpx::exception const& e);
-        inline explicit error_code(boost::exception_ptr const& e);
+        HPX_EXPORT error_code(int err, hpx::exception const& e);
+        HPX_EXPORT explicit error_code(boost::exception_ptr const& e);
 
         boost::exception_ptr exception_;
     };
@@ -284,102 +284,6 @@ namespace hpx
     {
         return error_code(mode);
     }
-
-    ///////////////////////////////////////////////////////////////////////////
-    // \cond NOINTERNAL
-    inline error_code::error_code(error e, throwmode mode)
-      : boost::system::error_code(make_system_error_code(e, mode))
-    {
-        if (e != success && e != no_success && !(mode & lightweight))
-            exception_ = detail::get_exception(e, "", mode);
-    }
-
-    inline error_code::error_code(error e, char const* func,
-            char const* file, long line, throwmode mode)
-      : boost::system::error_code(make_system_error_code(e, mode))
-    {
-        if (e != success && e != no_success && !(mode & lightweight)) {
-            exception_ = detail::get_exception(e, "", mode, func, file, line);
-        }
-    }
-
-    inline error_code::error_code(error e, char const* msg, throwmode mode)
-      : boost::system::error_code(make_system_error_code(e, mode))
-    {
-        if (e != success && e != no_success && !(mode & lightweight))
-            exception_ = detail::get_exception(e, msg, mode);
-    }
-
-    inline error_code::error_code(error e, char const* msg,
-            char const* func, char const* file, long line, throwmode mode)
-      : boost::system::error_code(make_system_error_code(e, mode))
-    {
-        if (e != success && e != no_success && !(mode & lightweight)) {
-            exception_ = detail::get_exception(e, msg, mode, func, file, line);
-        }
-    }
-
-    inline error_code::error_code(error e, std::string const& msg,
-            throwmode mode)
-      : boost::system::error_code(make_system_error_code(e, mode))
-    {
-        if (e != success && e != no_success && !(mode & lightweight))
-            exception_ = detail::get_exception(e, msg, mode);
-    }
-
-    inline error_code::error_code(error e, std::string const& msg,
-            char const* func, char const* file, long line, throwmode mode)
-      : boost::system::error_code(make_system_error_code(e, mode))
-    {
-        if (e != success && e != no_success && !(mode & lightweight)) {
-            exception_ = detail::get_exception(e, msg, mode, func, file, line);
-        }
-    }
-
-    inline error_code::error_code(int err, hpx::exception const& e)
-    {
-        this->boost::system::error_code::assign(err, get_hpx_category());
-        exception_ = get_exception_ptr(e);
-    }
-
-    inline error_code::error_code(boost::exception_ptr const& e)
-      : boost::system::error_code(make_system_error_code(get_error(e), rethrow)),
-        exception_(e)
-    {}
-
-    ///////////////////////////////////////////////////////////////////////////
-    inline std::string error_code::get_message() const
-    {
-        if (exception_) {
-            try {
-                boost::rethrow_exception(exception_);
-            }
-            catch (boost::exception const& be) {
-                return dynamic_cast<std::exception const*>(&be)->what();
-            }
-        }
-        return get_error_what(*this);   // provide at least minimal error text
-    }
-
-    ///////////////////////////////////////////////////////////////////////////
-    inline error_code& error_code::operator=(error_code const& rhs)
-    {
-        if (this != &rhs) {
-            if (rhs.value() == success) {
-                // if the rhs is a success code, we maintain our throw mode
-                this->boost::system::error_code::operator=(
-                    make_success_code(
-                        (category() == get_lightweight_hpx_category()) ?
-                            hpx::lightweight : hpx::plain));
-            }
-            else {
-                this->boost::system::error_code::operator=(rhs);
-            }
-            exception_ = rhs.exception_;
-        }
-        return *this;
-    }
-    /// \endcond
 }
 
 #include <hpx/throw_exception.hpp>
